+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/TeluguViewModel.java
@@ -9,6 +9,7 @@ import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.sdzshn3.android.news247.Fragments.TeluguNewsFragment;
+import com.sdzshn3.android.news247.Repositories.TeluguRepository;
 import com.sdzshn3.android.news247.TeluguNewsModel;
 import com.sdzshn3.android.news247.QueryUtils;
 import com.sdzshn3.android.news247.SupportClasses.DataHolder;
@@ -16,34 +17,14 @@ import com.sdzshn3.android.news247.SupportClasses.DataHolder;
 import java.util.List;
 
 public class TeluguViewModel extends AndroidViewModel {
-    private static MutableLiveData<List<TeluguNewsModel>> data = new MutableLiveData<>();
+    private TeluguRepository repository;
 
     public TeluguViewModel(@NonNull Application application) {
         super(application);
-        loadData();
-    }
-
-    public static void loadData() {
-        new AsyncTask<Void, Void, List<TeluguNewsModel>>() {
-            @Override
-            protected List<TeluguNewsModel> doInBackground(Void... voids) {
-                QueryUtils queryUtils = new QueryUtils();
-                try {
-                    return queryUtils.extractTeluguNewsFromRss(DataHolder.TELUGU_NEWS_REQUEST_URL, Integer.parseInt(TeluguNewsFragment.numberOfArticles));
-                } catch (NumberFormatException e) {
-                    Log.e("TeluguViewModel", "numberOfArticles not initialised");
-                    return null;
-                }
-            }
-
-            @Override
-            protected void onPostExecute(List<TeluguNewsModel> newsList) {
-                data.postValue(newsList);
-            }
-        }.execute();
+        repository = new TeluguRepository();
     }
 
     public LiveData<List<TeluguNewsModel>> getData() {
-        return data;
+        return repository.getData();
     }
 }
